listener/requirements.txt
pika==1.3.2
boto3==1.28.63
pillow==10.0.1
requests==2.31.0
psycopg2-binary==2.9.9
